requirements.txt
annotated-types==0.7.0
astroid==2.15.8
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
ffmpeg==1.4
ffmpeg-downloader==0.3.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.7
lazy-object-proxy==1.10.0
packaging==24.1
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyparsing==3.1.2
requests==2.32.3
requirements-detector==1.2.2
rsa==4.9
semver==3.0.2
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wrapt==1.16.0
pymediainfo
